Confirmatios ask, some css

+++ src/pages/AdvertPage/advertPage.tsx
@@ -3,14 +3,33 @@ import { IAds } from '../../interfaces/interfaces.ts';
 import Layout from '../../components/layout/Layout.tsx';
 import {getAd, deleteAd} from './service.ts';
 import { Link, useParams } from 'react-router-dom';
+import { useConfirm } from '../../context/confirmationContext/confirmCustomHook.ts';
 import './advertPage.css';
 import Button from '../../components/shared/Button.tsx';
 export default function AdvertPage() {
     const [ad, setAd] = useState<IAds>();
+    const {confirmState, onUnhidden, onDeleter} =useConfirm()
+    const [gonnaDelete, setGonnaDelete] = useState<boolean>();
     const params = useParams();
+    console.log(params)
+console.log(confirmState)
     const handleDeleteAd = ()=>{
-        params.adId && deleteAd(params.adId)
+        onDeleter()
+        onUnhidden()
+        setGonnaDelete(true)
+        console.log(gonnaDelete, confirmState)
+
     }
+    useEffect(() => {
+        console.log('message')
+        console.log(gonnaDelete, confirmState)
+        if( gonnaDelete && confirmState){
+           params.adId && deleteAd(params.adId)
+        }
+        setGonnaDelete(false)
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [confirmState]);
+
     useEffect(() => {
         const getDatad = async () => {
             if (params.adId) {
